feat: split balance and table details into two components
